+++ src/main/java/fontEditor/TileEditor.java
@@ -170,10 +170,6 @@ class TileEditor extends JPanel implements java.awt.event.MouseListener, java.aw
         font.setTilePixel(selectedTile, x, y, color);
     }
 
-    int getDirectPixel(int x, int y) {
-        return font.getPixel(x, y);
-    }
-
     public void mouseEntered(java.awt.event.MouseEvent e) {
     }